fix: ActiveButton 컴포넌트 모달 관련 수정
