+++ weekly-mission/next-project/src/components/FolderContent/index.tsx
@@ -3,6 +3,8 @@ import LinksContent from "./LinksContent";
 import { Folder } from "@api/useFoldersByUserId";
 import { FolderId } from "@components/Folder";
 import { LinkData } from "@api/fetchFolderLinksData";
+import Modal from "../Modal";
+import { useModal } from "src/context/ModalContext";
 
 export interface FolderContentProps {
   foldersData: Folder;
